+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
-- Block backspace on repl input.
+ - Fix `eval defun at cursor` action error. #121
+ - Block backspace on repl input.
 
 ## 2.1.0