+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fix `eval defun at cursor` action error. #121
  - Create view error on test error. #128
  - Block backspace on repl input.
+ - Update repl window ns after switching ns.
 
 ## 2.1.0