+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## 2.2.0
+
  - Fix `eval defun at cursor` action error. #121
  - Create view error on test error. #128
  - Block backspace on repl input.